[FRD-82] Reading company name on experience (#68)
## [FRD-82] Description
Introduces CompaniesWidget.scss for styling the CompaniesWidget component and imports it into the global stylesheet. Updates WidgetExperienceObject to optionally include a company, and modifies the experience widget column config to display the company name if available.

Changed WidgetCompany from a default export to a named export in CompaniesWidget.types.ts. Updated all relevant imports in CompaniesWidget.tsx and ExperiencesWidget.types.ts to use the named export for consistency and clarity.